+++ packages/cloud-cognitive/src/components/EditInPlace/EditInPlaceV2.stories.js
@@ -11,35 +11,19 @@ import {
   prepareStory,
 } from '../../global/js/utils/story-helper';
 import { action } from '@storybook/addon-actions';
-import { InlineEdit } from '../InlineEdit/InlineEdit';
-import { InlineEditV2 } from '.';
-import { DisplayBox } from '../../global/js/utils/DisplayBox';
-import mdx from './InlineEditV2.mdx';
-import styles from './_storybook-styles.scss';
-
-const storyClass = 'inline-edit-v2-example';
+import { EditInPlace } from '.';
+import mdx from '../InlineEditV2/InlineEditV2.mdx';
+import styles from '../InlineEditV2/_storybook-styles.scss';
 
 export default {
-  title: getStoryTitle(InlineEditV2.displayName),
-  component: InlineEditV2,
-  argTypes: {
-    containerWidth: {
-      control: { type: 'range', min: 20, max: 800, step: 10 },
-      description:
-        'Controls containing element width. Used for demonstration purposes, not property of the component.',
-    },
-  },
+  title: getStoryTitle(EditInPlace.displayName),
+  component: EditInPlace,
   parameters: {
     styles,
     docs: {
       page: mdx,
     },
   },
-  decorators: [
-    (story) => (
-      <DisplayBox className={`${storyClass}__viewport`}>{story()}</DisplayBox>
-    ),
-  ],
 };
 
 const actionSave = action('save');
@@ -48,11 +32,10 @@ const actionCancel = action('cancel');
 
 const defaultProps = {
   cancelLabel: 'Cancel',
-  containerWidth: 300,
   editLabel: 'Edit',
   id: 'story-id',
   invalid: false,
-  invalidText: 'This field is required',
+  invalidLabel: 'This field is required',
   labelText: 'Label text',
   onCancel: () => {},
   onChange: () => {},
@@ -64,7 +47,7 @@ const defaultProps = {
   value: 'default',
 };
 
-const Template = ({ containerWidth, ...args }) => {
+const Template = (args) => {
   const [value, setValue] = useState(defaultProps.value);
 
   const onChange = (val) => {
@@ -90,23 +73,12 @@ const Template = ({ containerWidth, ...args }) => {
     onCancel,
   };
 
-  return (
-    <div style={{ width: containerWidth }}>
-      <InlineEdit {...props} className="inline-edit-v2-example" />
-    </div>
-  );
+  return <EditInPlace {...props} className="inline-edit-v2-example" />;
 };
 
-export const Default = prepareStory(Template, {
+export const Version2 = prepareStory(Template, {
   args: {
     ...defaultProps,
-  },
-});
-
-export const Invalid = prepareStory(Template, {
-  args: {
-    ...defaultProps,
-    invalid: true,
   },
 });