+++ packages/cloud-cognitive/src/settings.js
@@ -4,6 +4,7 @@ import pkgSettings from './global/js/package-settings';
 import { settings as carbonSettings } from 'carbon-components';
 import React from 'react';
 import { themes } from '@carbon/themes';
+import pconsole from './global/js/utils/pconsole';
 
 export const carbon = {
   get prefix() {
@@ -17,6 +18,20 @@ export const carbon = {
   },
 };
 
+const componentDeprecatedWarning = (name, details) =>
+  `Carbon for IBM Products (WARNING): Component "${name}" is deprecated. ${details}`;
+
+pkgSettings.logDeprecated = (component, name) => {
+  if (component?.deprecated) {
+    const { level, details } = component.deprecated;
+    const logUsing = pconsole?.[level] ?? pconsole.error;
+
+    logUsing(
+      componentDeprecatedWarning(name || component.displayName, details)
+    );
+  }
+};
+
 // Check that a component is enabled. This function returns a stub which checks
 // the component status on first use and then renders as the component or as
 // a Canary placeholder initialized with the name of the replaced component.
@@ -25,9 +40,10 @@ export const carbon = {
 pkgSettings.checkComponentEnabled = (component, name) => {
   if (component.render) {
     // The component is a forward-ref, so make a stub forward-ref.
-    const forward = React.forwardRef((props, ref) =>
+    const forward = React.forwardRef((props, ref) => {
+      pkgSettings.logDeprecated(component, name); // may log don't care about result
       // Replace the stub's render fn so this test only happens once.
-      (forward.render =
+      return (forward.render =
         pkgSettings.isComponentEnabled(name) ||
         !pkgSettings.isComponentPublic(name)
           ? // If the component is enabled, or if it's not a public component,
@@ -39,8 +55,8 @@ pkgSettings.checkComponentEnabled = (component, name) => {
         // Call it now (after this it will be directly called).
         props,
         ref
-      )
-    );
+      );
+    });
 
     // Transfer object properties already assigned (eg propTypes, displayName)
     // then merge in the stub forward-ref which checks the component status
@@ -48,9 +64,10 @@ pkgSettings.checkComponentEnabled = (component, name) => {
     return Object.assign({}, component, forward);
   } else {
     // The component is a direct render fn, so make a stub render fn.
-    let render = (props) =>
+    let render = (props) => {
+      pkgSettings.logDeprecated(component, name); // may log don't care about result
       // Replace the stub render fn so this test only happens once.
-      (render =
+      return (render =
         pkgSettings.isComponentEnabled(name) ||
         !pkgSettings.isComponentPublic(name)
           ? // If the component is enabled, or if it's not a public component,
@@ -62,7 +79,7 @@ pkgSettings.checkComponentEnabled = (component, name) => {
         // Call it now (after this it will be directly called).
         props
       );
-
+    };
     // Transfer object properties already assigned (eg propTypes, displayName)
     // to a function which calls the stub render fn which checks the component
     // status when first used.